backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic[email]==2.4.2
pydantic-settings==2.0.3

# Environment management
python-decouple==3.8

# Payments
stripe==7.8.0

# Email
aiosmtplib==2.0.2
emails==0.6

# Utilities
python-dateutil==2.8.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.10.1
isort==5.12.0
mypy==1.7.0
